- clean-up work around the codebase  * removed <windows.h> include from the global namespace (now only included if absolutely necessary)  * replaced some Windows-specific types with standard types for abstraction - split image loading logic from tex add dialog for better management and re-usage
preparing to work on the builder...

+++ rwlib/src/txdread.d3d9.cpp
@@ -258,7 +258,7 @@ void d3d9NativeTextureTypeProvider::DeserializeTexture( TextureBase *theTexture,
                             // If the user wants to know about such things, notify him.
                             if ( engineIgnoreSecureWarnings == false )
                             {
-                                engineInterface->PushWarning( "texture " + theTexture->GetName() + " has an unknown D3DFORMAT link (" + std::to_string( (DWORD)d3dFormat ) + ")" );
+                                engineInterface->PushWarning( "texture " + theTexture->GetName() + " has an unknown D3DFORMAT link (" + std::to_string( d3dFormat ) + ")" );
 
                                 hasReportedStrongWarning = true;
                             }
